refactor/LJJ(#137): 채팅 폴더 구조 변경
