+++ src/components/Blog/privateBlog/PrivateBlogContents.tsx
@@ -3,7 +3,7 @@ import SelectBlogDateSection from './SelectBlogDateSection';
 import PrivateBlogList from './PrivateBlogList';
 import { Suspense } from 'react';
 import { Loading } from '@/components';
-import ChatBox from '@/components/Chat/ChatBox/ChatBox';
+import ChatBox from '@/components/Chat/ChatBox';
 
 export default function PrivateBlogContents() {
   return (
