+++ src/components/Chat/ChatBox.tsx
@@ -1,12 +1,13 @@
 import * as S from './style';
 import { Svg } from '@/components/Svg';
 import useLocalStorage from '@/hook/useLocalStorage';
-import ChatRoomModal from './ChatModal';
+import ChatRoomModal from './ChatBox/ChatModal';
 import { ChatQueries } from '@/apis/chat';
-import { useSocket } from '@/hook/useSocket';
 import { useEffect, useRef, useState } from 'react';
-import { scrollToBottom, useReceivedMessage } from '../hooks';
 import { changeInfo } from '@/utils';
+import { scrollToBottom } from './utils';
+import { useMessageListener } from './hooks';
+import { useSocket } from '@/hook/useSocket';
 
 export default function ChatBox() {
   const [isOpenChatModal, setIsOpenChatModal] = useState({
@@ -15,31 +16,19 @@ export default function ChatBox() {
   const { value: token } = useLocalStorage('accessToken');
   const belongToChatRoom = ChatQueries.useGetBelongToChatRoom();
 
-  const socket = useSocket();
-  const { onMessageReceived } = useReceivedMessage();
-
   const scrollToBottomRef = useRef<HTMLDivElement>(null);
+  const socket = useSocket();
 
   const handlerOpenChatModal = changeInfo.toggle({
     setState: setIsOpenChatModal,
     key: 'chat',
   });
 
-  useEffect(() => {
-    socket?.emit('enter_chat_room', { roomId: belongToChatRoom?.id });
-    socket?.on('receive_message', res => {
-      onMessageReceived(res);
-      scrollToBottom(scrollToBottomRef);
-    });
-
-    return () => {
-      socket?.off('enter_chat_room');
-      socket?.off('receive_message', res => {
-        onMessageReceived(res);
-        scrollToBottom(scrollToBottomRef);
-      });
-    };
-  }, [socket]);
+  useMessageListener({
+    roomId: belongToChatRoom?.id,
+    scrollToBottomRef,
+    socket,
+  });
 
   useEffect(() => {
     if (isOpenChatModal) {
