+++ src/components/Chat/ChatBox/style.ts
@@ -100,24 +100,6 @@ export const SendIconButton = styled.button`
   margin: 3px;
 `;
 
-export const ButtonWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50%;
-  padding: 8px;
-  border: 1px solid ${({ theme }) => theme.button.primary.base};
-  background-color: ${({ theme }) => theme.button.primary.base};
-  position: fixed;
-  right: 5%;
-  bottom: 5%;
-  z-index: 88;
-  cursor: pointer;
-  &:hover {
-    background-color: ${({ theme }) => theme.button.primary.hover};
-  }
-`;
-
 export const ChatMessage = styled.div`
   display: flex;
   flex-direction: column;
