+++ src/components/Chat/ChatBox/type.d.ts
@@ -1,7 +1,4 @@
-import {
-  BelongToMeChatRoomReturnType,
-  ChatMessagePaginationType,
-} from '@/apis/chat/type';
+import { BelongToMeChatRoomReturnType } from '@/apis/chat/type';
 import { Socket } from 'socket.io-client';
 
 export interface ChatModalProps {
@@ -11,25 +8,10 @@ export interface ChatModalProps {
   scrollToBottomRef: React.RefObject<HTMLDivElement>;
 }
 
-export interface ChatMessageListType {
-  id: string;
-  createdAt: string;
-  updatedAt: string;
-  roomId: string;
-  senderId: string;
-  message: string;
-  blogPostUrl: string | null;
-}
-
 export interface ChatServerBaseResponse {
   message: string;
   roomId: string;
   blogPostUrl: string;
-}
-
-export interface ChatCurrentDataType {
-  pageParam: Array;
-  pages: ChatMessagePaginationType[];
 }
 
 export interface ChatAckResponse {
