+++ src/components/Chat/utils.ts
@@ -1,0 +1,8 @@
+export const scrollToBottom = (
+  scrollToBottomRef: React.RefObject<HTMLDivElement>
+) => {
+  if (scrollToBottomRef.current) {
+    scrollToBottomRef.current.scrollTop =
+      scrollToBottomRef.current.scrollHeight;
+  }
+};
